ONLINE_MODULE: performance 'n display improvements
If the numonline, which i renamed to numguild to be coherent with the
variable names, is empty, then don't even call the functions at all, as
they will return an empty $list as well and just waste system calls.

If there are only people in the priv group, don't display the guild
header at all, and give the private group the main header instead.